requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
numpy
matplotlib
pydantic